Added multi-element add/del, refreshed theme
With the new theme comes a dark theme option and an icon switch.
The multi-element operation is a first step toward more complex
operations such as moving and copying one or more objects.

I also switched back to the MaraiDB python library as it is no longer in
alpha and they've fixed one major breaking bug which made it impossible
to insert negative integers.